pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Zevobla <ghost-brake-your-home@protonmail.com>"]
license = "AGPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
fastapi = "^0.110.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
nuitka = "^2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
